pyproject.toml
[project]
name = "scalable-user-service"
version = "1.0.0"
description = "Production-grade user authentication and management microservice"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "email-validator>=2.0.0",
    "fastapi>=0.136.0",
    "pydantic-settings>=2.0.0",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.45.0",
    "passlib[bcrypt]>=1.7.4",
    "greenlet>=3.4.0",
    "bcrypt<5",
    "python-jose[cryptography]>=3.3.0",
    "slowapi>=0.1.9",
    "structlog>=25.5.0",
    "resend>=2.29.0",
    "prometheus-client>=0.25.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "locust>=2.43.4",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

[tool.ruff]
target-version = "py313"
line-length = 120
exclude = ["alembic/"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501", "B008", "B904"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
